Eliminate compile warnings on Windows 64 bit. (~2800 non-unique from falcon)
 -Most of the warnings is about size_t to int truncation in memory manager.
Changed MemMgr function prototypes to accept size_t lengths.
 alloc() will now assert if the size is more then 2GB.
 (falcon cannot allocate more at once because it tracks size of each 
allocation with an "int")

- the rest of the warnings seems harmless and  solved mostly with casts.

+++ storage/falcon/Cache.cpp
@@ -852,7 +852,7 @@ void Cache::ioThread(void)
 						
 					flushLock.unlock();
 					//Log::debug(" %d Writing %s %d pages: %d - %d\n", thread->threadId, (const char*) dbb->fileName, count, pageNumber, pageNumber + count - 1);
-					int length = p - buffer;
+					int length = (int)(p - buffer);
 					priority.schedule(PRIORITY_LOW);
 					
 					try
